# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PlayGame Gateway API",
    "description": "Gateway de Pagamentos PIX - Solução completa para integração de pagamentos instantâneos",
    "version": "1.0.0",
    "contact": {
      "name": "PlayGame Gateway",
      "email": "redacted@example.invalid",
      "url": "https://playgameoficial.com.br"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://playgameoficial.com.br/api",
      "description": "Servidor de Produção"
    },
    {
      "url": "https://sandbox.playgameoficial.com.br/api",
      "description": "Servidor de Sandbox"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/wallet/deposit/payment": {
      "post": {
        "tags": ["PIX"],
        "summary": "Gerar QR Code PIX",
        "description": "Cria uma transação de depósito PIX e retorna o QR Code para pagamento",
        "operationId": "createPixDeposit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PixDepositRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Depósito Básico",
                  "value": {
                    "token": "seu_token_aqui",
                    "secret": "sua_chave_secreta",
                    "amount": 100.00,
                    "client_name": "João Silva",
                    "client_email": "redacted@example.invalid",
                    "client_document": "12345678901",
                    "client_telefone": "11999999999",
                    "description": "Depósito via PIX"
                  }
                },
                "with_split": {
                  "summary": "Depósito com Split",
                  "value": {
                    "token": "seu_token_aqui",
                    "secret": "sua_chave_secreta",
                    "amount": 1000.00,
                    "client_name": "João Silva",
                    "client_email": "redacted@example.invalid",
                    "client_document": "12345678901",
                    "client_telefone": "11999999999",
                    "description": "Depósito via PIX",
                    "split_email": "redacted@example.invalid",
                    "split_percentage": 10.0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixDepositResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pixout": {
      "post": {
        "tags": ["PIX"],
        "summary": "Fazer um Pagamento PIX",
        "description": "Processa um saque PIX para a conta do cliente",
        "operationId": "createPixWithdraw",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PixWithdrawRequest"
              },
              "examples": {
                "phone": {
                  "summary": "Saque para Telefone",
                  "value": {
                    "token": "seu_token_aqui",
                    "secret": "sua_chave_secreta",
                    "amount": 50.00,
                    "pixKey": "11999999999",
                    "pixKeyType": "phone",
                    "description": "Saque via PIX"
                  }
                },
                "email": {
                  "summary": "Saque para Email",
                  "value": {
                    "token": "seu_token_aqui",
                    "secret": "sua_chave_secreta",
                    "amount": 100.00,
                    "pixKey": "redacted@example.invalid",
                    "pixKeyType": "email",
                    "description": "Saque via PIX"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixWithdrawResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "IP não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "post": {
        "tags": ["PIX"],
        "summary": "Consultar Transação",
        "description": "Consulta o status de uma transação específica",
        "operationId": "getTransactionStatus",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              },
              "examples": {
                "check_status": {
                  "summary": "Verificar Status",
                  "value": {
                    "idTransaction": "dep_1234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Transação não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/balance": {
      "post": {
        "tags": ["Account"],
        "summary": "Consultar Saldo",
        "description": "Consultar o saldo disponível da conta",
        "operationId": "getBalance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "PixDepositRequest": {
        "type": "object",
        "required": ["token", "secret", "amount", "client_name", "client_email", "client_document"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Token de autenticação",
            "example": "seu_token_aqui"
          },
          "secret": {
            "type": "string",
            "description": "Chave secreta",
            "example": "sua_chave_secreta"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "minimum": 1.00,
            "maximum": 50000.00,
            "description": "Valor do depósito em reais",
            "example": 100.00
          },
          "client_name": {
            "type": "string",
            "description": "Nome do cliente",
            "example": "João Silva"
          },
          "client_email": {
            "type": "string",
            "format": "email",
            "description": "Email do cliente",
            "example": "redacted@example.invalid"
          },
          "client_document": {
            "type": "string",
            "description": "CPF/CNPJ do cliente",
            "example": "12345678901"
          },
          "client_telefone": {
            "type": "string",
            "description": "Telefone do cliente",
            "example": "11999999999"
          },
          "description": {
            "type": "string",
            "description": "Descrição da transação",
            "example": "Depósito via PIX"
          },
          "split_email": {
            "type": "string",
            "format": "email",
            "description": "Email do destinatário do split",
            "example": "redacted@example.invalid"
          },
          "split_percentage": {
            "type": "number",
            "format": "decimal",
            "minimum": 0,
            "maximum": 100,
            "description": "Percentual do split",
            "example": 10.0
          }
        }
      },
      "PixDepositResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "idTransaction": {
                "type": "string",
                "description": "ID único da transação",
                "example": "dep_1234567890"
              },
              "status": {
                "type": "string",
                "enum": ["pending", "paid", "failed", "cancelled"],
                "example": "pending"
              },
              "amount": {
                "type": "number",
                "format": "decimal",
                "example": 100.00
              },
              "qr_code": {
                "type": "string",
                "description": "QR Code em base64",
                "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
              },
              "qr_code_text": {
                "type": "string",
                "description": "Código PIX para copiar e colar",
                "example": "00020126580014br.gov.bcb.pix..."
              },
              "expires_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-15T15:30:00Z"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-15T15:00:00Z"
              }
            }
          }
        }
      },
      "PixWithdrawRequest": {
        "type": "object",
        "required": ["token", "secret", "amount", "pixKey", "pixKeyType"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Token de autenticação",
            "example": "seu_token_aqui"
          },
          "secret": {
            "type": "string",
            "description": "Chave secreta",
            "example": "sua_chave_secreta"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "minimum": 10.00,
            "maximum": 20000.00,
            "description": "Valor do saque em reais",
            "example": 50.00
          },
          "pixKey": {
            "type": "string",
            "description": "Chave PIX de destino",
            "example": "11999999999"
          },
          "pixKeyType": {
            "type": "string",
            "enum": ["cpf", "cnpj", "email", "phone", "random"],
            "description": "Tipo da chave PIX",
            "example": "phone"
          },
          "description": {
            "type": "string",
            "description": "Descrição da transação",
            "example": "Saque via PIX"
          }
        }
      },
      "PixWithdrawResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "idTransaction": {
                "type": "string",
                "description": "ID único da transação",
                "example": "out_1234567890"
              },
              "status": {
                "type": "string",
                "enum": ["processing", "completed", "failed", "cancelled"],
                "example": "processing"
              },
              "amount": {
                "type": "number",
                "format": "decimal",
                "example": 50.00
              },
              "pixKey": {
                "type": "string",
                "example": "11999999999"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-15T15:00:00Z"
              }
            }
          }
        }
      },
      "StatusRequest": {
        "type": "object",
        "required": ["idTransaction"],
        "properties": {
          "idTransaction": {
            "type": "string",
            "description": "ID da transação",
            "example": "dep_1234567890"
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "idTransaction": {
                "type": "string",
                "example": "dep_1234567890"
              },
              "status": {
                "type": "string",
                "enum": ["pending", "paid", "processing", "completed", "failed", "cancelled"],
                "example": "paid"
              },
              "amount": {
                "type": "number",
                "format": "decimal",
                "example": 100.00
              },
              "paid_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-15T15:05:00Z"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-15T15:00:00Z"
              }
            }
          }
        }
      },
      "BalanceRequest": {
        "type": "object",
        "required": ["token", "secret"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Token de autenticação",
            "example": "seu_token_aqui"
          },
          "secret": {
            "type": "string",
            "description": "Chave secreta",
            "example": "sua_chave_secreta"
          }
        }
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "balance": {
                "type": "number",
                "format": "decimal",
                "description": "Saldo total",
                "example": 1500.00
              },
              "available_balance": {
                "type": "number",
                "format": "decimal",
                "description": "Saldo disponível",
                "example": 1200.00
              },
              "pending_balance": {
                "type": "number",
                "format": "decimal",
                "description": "Saldo pendente",
                "example": 300.00
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro",
            "example": "Dados inválidos"
          },
          "client_ip": {
            "type": "string",
            "description": "IP do cliente (quando aplicável)",
            "example": "192.168.1.1"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "PIX",
      "description": "Operações PIX (Depósitos e Saques)"
    },
    {
      "name": "Account",
      "description": "Operações de Conta"
    },
    {
      "name": "Webhooks",
      "description": "Notificações em tempo real"
    }
  ]
}